Miscellaneous cleanup and refactoring (#41)
* Reduce code duplication in `build.rs`, define some configuration symbols

* Reorganize and clean up the `io` module a bit

* Clean up re-exports, correctly alias target with capitalized name

* Move `Transport` type into library, extract some functions

* Change type of `tinfl_decompressor` struct's `bit_buf` field from u64 to u32

* We can now build using `opt-level = "z"`!

* Update `README.md`, rustfmt

* Tweak symbol name for RISC-V in build script

+++ src/miniz_types.rs
@@ -43,7 +43,7 @@ pub struct tinfl_decompressor {
     counter: u32,
     num_extra: u32,
     table_sizes: [u32; TINFL_MAX_HUFF_TABLES],
-    bit_buf: u64,
+    bit_buf: u32,
     dist_from_out_buf_start: u32,
     tables: [tinfl_huff_table; TINFL_MAX_HUFF_TABLES],
     raw_header: [u8; 4],
